Renamed files to better represent usage. Fixed import conflicts. Consolidated Config and Backend. Files are now saved on a new folder. Deleted Superfluous files.
