+++ .gitignore
@@ -1,4 +1,2 @@
-name.txt
-encryptedkey.txt
 __pycache__
 ```
