--- Config/accounts.json
@@ -1,12 +1,0 @@
-[
-    {
-        "id": 1,
-        "websiteName": "Youtube",
-        "username": "Enrei",
-        "password": "ActualPasswordNotJustMyLastName",
-        "dateCreated": "2023-04-12",
-        "dateModified": "2023-04-12",
-        "lastChecked": "2023-04-12",
-        "lastModified": "2023-04-12"
-    }
-]
